requirements.txt
streamlit
PyPDF2
python-docx
pytesseract
Pillow
requests
